+++ app/src/main/java/com/example/networksocial/Fragments/HomeFragment.java
@@ -3,6 +3,7 @@ package com.example.networksocial.Fragments;
 import android.content.Intent;
 import android.os.Bundle;
 
+import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
@@ -12,15 +13,23 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Button;
 
 import com.example.networksocial.Activities.DashboardActivity;
 import com.example.networksocial.Activities.MainActivity;
 import com.example.networksocial.R;
+import com.google.android.gms.auth.api.signin.GoogleSignIn;
+import com.google.android.gms.auth.api.signin.GoogleSignInClient;
+import com.google.android.gms.auth.api.signin.GoogleSignInOptions;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 
 public class HomeFragment extends Fragment {
-
+    GoogleSignInOptions gso;
+    GoogleSignInClient gsc;
+    private GoogleSignInClient mGoogleSignInClient;
     FirebaseAuth firebaseAuth;
 
     public HomeFragment() {
@@ -32,6 +41,20 @@ public class HomeFragment extends Fragment {
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
         View view =  inflater.inflate(R.layout.fragment_home, container, false);
+
+        initGoogleSignInClient();
+
+        Button button = view.findViewById(R.id.btn_SignOut);
+
+        // Gắn sự kiện onClick cho button
+        button.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                // Xử lý sự kiện onClick ở đây
+                signOut();
+            }
+        });
+
         firebaseAuth = FirebaseAuth.getInstance();
         return view;
     }
@@ -72,4 +95,28 @@ public class HomeFragment extends Fragment {
         }
         return super.onOptionsItemSelected(item);
     }
+
+    private void signOut() {
+        mGoogleSignInClient.signOut()
+                .addOnCompleteListener(getActivity(), new OnCompleteListener<Void>() {
+                    @Override
+                    public void onComplete(@NonNull Task<Void> task) {
+                        // Xử lý khi đăng xuất thành công
+                        // Ví dụ: Chuyển đến màn hình đăng nhập lại
+                        Intent intent = new Intent(getContext(), MainActivity.class);
+
+                        startActivity(intent);
+                    }
+                });
+    }
+
+    // Khởi tạo GoogleSignInClient trong phương thức onCreate hoặc onStart
+    private void initGoogleSignInClient() {
+        GoogleSignInOptions gso = new GoogleSignInOptions.Builder(GoogleSignInOptions.DEFAULT_SIGN_IN)
+                .requestEmail()
+                .build();
+
+        mGoogleSignInClient = GoogleSignIn.getClient(getActivity(), gso);
+    }
+
 }
